move more of the processmanger fixture
