+++ source/dotnet/Services/IntegrationTests/IntegrationTests.csproj
@@ -75,6 +75,7 @@ limitations under the License.
 
   <ItemGroup>
     <Folder Include="Fixtures\Mock" />
+    <Folder Include="Fixtures\TestCommon\Fixture\FunctionApp" />
     <Folder Include="ProcessManager" />
   </ItemGroup>
