+++ source/dotnet/Services/ProcessManager.IntegrationTests/Fixtures/ProcessManagerFunctionAppCollectionFixture.cs
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.IntegrationTests.Fixtures.TestCommon.Fixture.FunctionApp;
 using Xunit;
 
-namespace Energinet.DataHub.Wholesale.IntegrationTests.TestCommon.Fixture;
+namespace ProcessManager.IntegrationTests.Fixtures;
 
 /// <summary>
 /// An xUnit collection fixture for ensuring tests don't run in parallel.
